remove_volume_child: fix warning

+++ src/search.c
@@ -372,7 +372,7 @@ void remove_volume_child(volume_device_extension* vde, volume_child* vc, BOOL no
                             ERR("IOCTL_MOUNTDEV_QUERY_DEVICE_NAME returned %08x\n", Status);
                         else {
                             MOUNTDEV_NAME* mdn2;
-                            ULONG mdnsize = offsetof(MOUNTDEV_NAME, Name[0]) + mdn.NameLength;
+                            ULONG mdnsize = (ULONG)offsetof(MOUNTDEV_NAME, Name[0]) + mdn.NameLength;
 
                             mdn2 = ExAllocatePoolWithTag(PagedPool, mdnsize, ALLOC_TAG);
                             if (!mdn2)
